CHORE: restructure graph reducer to include utility function for update
